requirements-dev.txt
black==19.10b0
eth-brownie>=1.11.7,<2.0.0
